@@ -1,25 +1,27 @@
 {
-  "name": "my-app",
+  "name": "react-video-abr",
   "version": "0.1.0",
-  "private": true,
+  "license": "MIT",
+  "homepage": ".",
   "dependencies": {
-    "cloudinary-core": "^2.11.4",
-    "cloudinary-react": "^1.7.0",
-    "cloudinary-video-player": "^1.5.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "tar": ">=6.1.2",
-    "url-parse": ">=1.5.2",
-    "urlsafe-base64": "^1.0.0",
-    "video.js": ">=7.14.3",
-    "web-vitals": "^1.0.1"
-  },
-  "devDependencies": {
-    "react-scripts": "^4.0.3"
+    "@cloudinary/react": "^1.4.1",
+    "@cloudinary/url-gen": "^1.8.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "cloudinary": "^1.30.1",
+    "cloudinary-video-player": "^1.9.1",
+    "dotenv": "^16.0.1",
+    "gh-pages": "^4.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -28,7 +30,6 @@
       "react-app/jest"
     ]
   },
-  "homepage": "https://cloudinary-training.github.io/cld-advanced-concepts/video-player/react/build",
   "browserslist": {
     "production": [
       ">0.2%",
